requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data manipulation and analysis
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter environment
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Computer vision utilities
opencv-python>=4.5.0
scikit-image>=0.18.0

# Model evaluation and metrics
scikit-learn>=1.0.0

# Progress bars and utilities
tqdm>=4.62.0

# Configuration management
pyyaml>=5.4.0

# Optional: for advanced visualization
tensorboard>=2.7.0

# Optional: for model deployment
onnx>=1.10.0
